+++ core/src/dird/ua_select.h
@@ -1,7 +1,7 @@
 /*
    BAREOS® - Backup Archiving REcovery Open Sourced
 
-   Copyright (C) 2018-2021 Bareos GmbH & Co. KG
+   Copyright (C) 2018-2022 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -59,9 +59,9 @@ bool SelectClientDbr(UaContext* ua, ClientDbRecord* cr);
 void StartPrompt(UaContext* ua, const char* msg);
 void AddPrompt(UaContext* ua, const char* prompt);
 void AddPrompt(UaContext* ua, std::string&& prompt);
-std::string FormatMulticolumnPrompts(const UaContext* ua,
-                                     int window_width,
-                                     int min_lines_threshold);
+std::string FormatPrompts(const UaContext* ua,
+                          int window_width,
+                          int min_lines_threshold);
 int DoPrompt(UaContext* ua,
              const char* automsg,
              const char* msg,